Move vulcan-lib validation.js to server-side only
