+++ packages/vulcan-lib/lib/modules/index.js
@@ -26,6 +26,5 @@ export * from './errors.js';
 export * from './intl.js';
 export * from './detect_locale.js';
 export * from './graphql_templates.js';
-export * from './validation.js';
 export * from './handleOptions';
 // export * from './resolvers.js';
